requirements.txt
playwright
pytest
pytest-playwright
pytest-html
python-dotenv
